make sockaddr operator() work closer to c++ spec
the problem is, sockaddr_in is a POD and pxfe_sockaddr_in is not.  so
taking "(sockaddr *)this" is not guaranteed by c++ specs to return a
pointer to the beginning of the "sockaddr_in" base class of
pxfe_sockaddr. (it does on g++ on all version i know of, but that's
just luck.) however, "static_cast<sockaddr_in*>(this)" is guaranteed
to do so. therefore the proper way to get a "sockaddr*" is to
double-cast, once to guarantee a downcast to the POD base type, then a
c-style coersion to sockaddr*. same for sockaddr_un. also add a
"socklen_t sasize()" method so we're not relying on
sizeof(pxfe_sockaddr_in) == sizeof(sockaddr_in), which is also not
technically guaranteed (even though pxfe_sockaddr_in is not
polymorphic and adds no explicit storage).