fonts dont get built right if bdftopcf isnt installed
a "process1 | process2" in a makefile never works right if process1
fails -- this is because the exit status of the whole line is that of
process2 only. in this case, if bdftopcf is not installed, you still
get PCF.GZ files but they have only the gzip header/footer with zero
bytes of contents.  also, never just "command > target" in a makefile
rule because now the target exists and the makefile rule will never
run again. do it to a temp file name and then, if the command
succeeds, "mv" to the correct target name.